ai_models_server/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
tqdm==4.65.0
seaborn==0.12.2
scikit-learn==1.2.2
google-generativeai==0.3.0
python-dotenv==1.0.0
ultralytics==8.0.120
requests==2.31.0
